스테이터스 효과 관리시 int에서 float을 사용하도록 변경

+++ Assets/_Project/Scripts/Unidad/StatusManager.cs
@@ -115,12 +115,12 @@ public class StatusManager
     public AttackStatus AttackUnitModifier => attackUnitModifier;
     public DefenceStatus DefenceUnitModifier => defenceUnitModifier;
 
-    public Dictionary<IUnitModifier, int> UnitModifiers => unitModifiers;
+    public Dictionary<IUnitModifier, float> UnitModifiers => unitModifiers;
 
     private readonly AttackStatus attackUnitModifier = new();
     private readonly DefenceStatus defenceUnitModifier = new();
 
-    private readonly Dictionary<IUnitModifier, int> unitModifiers = new();
+    private readonly Dictionary<IUnitModifier, float> unitModifiers = new();
     #endregion
 
     #endregion
